requirements.txt
backoff==1.11.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Deprecated==1.2.14
googleapis-common-protos==1.65.0
grpcio==1.67.1
idna==3.10
importlib_metadata==8.5.0
intervaltree==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
opentelemetry-api==1.28.0
opentelemetry-exporter-otlp==1.28.0
opentelemetry-exporter-otlp-proto-common==1.28.0
opentelemetry-exporter-otlp-proto-grpc==1.28.0
opentelemetry-exporter-otlp-proto-http==1.28.0
opentelemetry-instrumentation==0.49b0
opentelemetry-instrumentation-wsgi==0.49b0
opentelemetry-proto==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-semantic-conventions==0.49b0
opentelemetry-util-http==0.49b0
packaging==24.1
protobuf==5.28.3
requests==2.32.3
sortedcontainers==2.4.0
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.2
wrapt==1.16.0
zipp==3.20.2
